+++ kitty/borders.py
@@ -7,6 +7,7 @@ from typing import Iterable, NamedTuple, Sequence
 from .fast_data_types import BORDERS_PROGRAM, add_borders_rect, get_options, init_borders_program, os_window_has_background_image
 from .shaders import program_for
 from .typing import LayoutType
+from .utils import color_as_int
 from .window_list import WindowGroup, WindowList
 
 
@@ -96,7 +97,7 @@ class Borders:
         active_group = all_windows.active_group
 
         for i, wg in enumerate(groups):
-            window_bg = wg.default_bg
+            window_bg = color_as_int(wg.default_bg)
             window_bg = (window_bg << 8) | BorderColor.window_bg
             if draw_borders and not draw_minimal_borders:
                 # Draw the border rectangles